int to Integer and long to Long

+++ src/test/java/com/shopify/ShopifySdkTest.java
@@ -1146,7 +1146,7 @@ public class ShopifySdkTest {
 				actualCreateRequestBody.getContent().get("product").get("variants").get(0).get("sku").asText());
 		assertEquals(somePrice,
 				actualCreateRequestBody.getContent().get("product").get("variants").get(0).get("price").decimalValue());
-		assertEquals(shopifyVariantCreationRequest.getRequest().getGrams(),
+		assertEquals(shopifyVariantCreationRequest.getRequest().getGrams().longValue(),
 				actualCreateRequestBody.getContent().get("product").get("variants").get(0).get("grams").asLong());
 		assertNull(actualCreateRequestBody.getContent().get("product").get("variants").get(0).get("available"));
 		assertEquals(shopifyVariantCreationRequest.getRequest().getOption1(),
@@ -1218,7 +1218,7 @@ public class ShopifySdkTest {
 		shopifyVariant.setSku("ABC-123");
 		shopifyVariant.setImageId("1");
 		shopifyVariant.setPrice(BigDecimal.valueOf(42.11));
-		shopifyVariant.setGrams(12);
+		shopifyVariant.setGrams(12L);
 		shopifyVariant.setInventoryQuantity(3L);
 		shopifyVariant.setRequiresShipping(true);
 		shopifyVariant.setTaxable(true);
@@ -1274,7 +1274,7 @@ public class ShopifySdkTest {
 				actualCreateRequestBody.getContent().get("product").get("variants").get(0).get("sku").asText());
 		assertEquals(shopifyProduct.getVariants().get(0).getPrice(),
 				actualCreateRequestBody.getContent().get("product").get("variants").get(0).get("price").decimalValue());
-		assertEquals(shopifyProduct.getVariants().get(0).getGrams(),
+		assertEquals(shopifyProduct.getVariants().get(0).getGrams().longValue(),
 				actualCreateRequestBody.getContent().get("product").get("variants").get(0).get("grams").asLong());
 		assertNull(actualCreateRequestBody.getContent().get("product").get("variants").get(0).get("available"));
 		assertEquals(shopifyProduct.getVariants().get(0).getOption1(),
@@ -1414,7 +1414,7 @@ public class ShopifySdkTest {
 		shopifyVariant.setSku("ABC-123");
 		shopifyVariant.setImageId("1");
 		shopifyVariant.setPrice(BigDecimal.valueOf(42.11));
-		shopifyVariant.setGrams(12);
+		shopifyVariant.setGrams(12L);
 		shopifyVariant.setInventoryQuantity(3L);
 		shopifyVariant.setRequiresShipping(true);
 		shopifyVariant.setTaxable(true);
@@ -1847,7 +1847,7 @@ public class ShopifySdkTest {
 		shopifyVariant.setSku("ABC-123");
 		shopifyVariant.setImageId("1");
 		shopifyVariant.setPrice(BigDecimal.valueOf(42.11));
-		shopifyVariant.setGrams(12);
+		shopifyVariant.setGrams(12L);
 		shopifyVariant.setInventoryQuantity(3L);
 		shopifyVariant.setRequiresShipping(true);
 		shopifyVariant.setTaxable(true);
